+++ contracts/isOdd.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.24;
 
 contract isOdd {
-    function isOddNumber(uint _number) public pure returns (bool) {
+    function isOddNumber(uint256 _number) public pure returns (bool) {
         return _number % 2 != 0;
     }
 }
